refactored to an object wrapper
